Run quick sanity check before spike and body collisions
Combines both the spike and body collision O(n^2) loops into one
loop which begins with a quick sanity check. No further actions will
be taken if the two bodies are separated by more than the length
of the longest spike.

This speeds up the simulation considerably, reducing CPU costs for that
component by about 70%, and ~30% for the app overall.

Signed-off-by: Zac Delventhal <delventhalz@gmail.com>