Remove mutation from genome module
Removing all mutation for the app is not practical (at least without
pulling in Immutable.js). It would require destroying and rebuilding
thousands of meebas and spikes, sixty times a second. In early tests,
it took garbage collector CPU share from ~1% to ~10%.

However, it is certainly possible to use _less_ mutation, and the
genome module is a perfect place to use some more declarative pure
code.

Signed-off-by: Zac Delventhal <delventhalz@gmail.com>