Use next location to check if bodies can interact
The canInteract check is somewhat problematic because the real
collision check is based on _both_ the bodies' current location and
next location. The canInteract check should be as simple as
possible, so it can only use one of those locations. Previously,
it used current location. This caused intermittent test failures,
and is probably on the whole slightly more buggy.

Signed-off-by: Zac Delventhal <delventhalz@gmail.com>