Use requestAnimationFrame for simulation
When the browser is out of focus, CPU usage is significantly throttled
to the point where it is not possible to complete even one simulation
frame per second. Setting a maximum simulation time per frame set will
prevent mass die off, but will still result in different selection
pressures when the browser is backgrounded. The simulation time is also
so slow, that it might as well be paused.

With the simulation loop running with requestAnimationFrame, it will
stop while the browser is in the background, pausing the simulation
and ensuring a consistent frame rate.

In the future it may be possible to use sweeping collision to decouple
frame rate from spike drain rate, allowing the simulation to run in the
background at 1Hz or less without so drastically affecting selection
pressure. For now, this will have to suffice.

Fixes #31

Signed-off-by: Zac Delventhal <delventhalz@gmail.com>