Fine-tune mass and spike settings
Current simulations run pretty well, but the meebas are a little
small, with a few too many spikes, that are a bit long. These tweaks
should result in a pretty enjoyable simulation at a variety of
screen sizes while allowing for a pretty diverse population of
meebas.

There is probably more work that could be done, but these will be
the final default settings for awhile.

Closes #30

Signed-off-by: Zac Delventhal <delventhalz@gmail.com>