Add property transform functions to getTweener
Transform objects can now include property transform functions.
These take the delta, the base value, the key, and the target object
being mutated, and should return the value to set to that key.

There is no immediate need for this feature, but it will be necessary
in the future when nested transform objects are introduced.

Includes a couple of helper methods which are used internally and
will also be useful for building common tween functions.

Signed-off-by: Zac Delventhal <delventhalz@gmail.com>